refactor: relocate StateWindowProvider to homeassistant package
StateWindowProvider was domain-misfiled. It consumes HA state-change
events (HandleStateChange matches homeassistant.StateWatchHandler), so
HA owns its data path. Keeping it in awareness/ forced every HA-adjacent
caller to reach into awareness for what should be a local type.

Pure relocation: package declaration changes awareness -> homeassistant,
the single instantiation site in app/new_awareness.go swaps to the new
import (already present for state watcher wiring). No logic changes.

Refs #625.
